[KYV-739] Made review fixes.
