+++ orderapi/src/main/java/fi/hel/verkkokauppa/order/model/FlowStep.java
@@ -1,6 +1,7 @@
 package fi.hel.verkkokauppa.order.model;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
@@ -10,6 +11,7 @@ import org.springframework.data.elasticsearch.annotations.FieldType;
 @Document(indexName = "flowsteps")
 @Setter
 @Getter
+@NoArgsConstructor
 public class FlowStep {
 
     @Id
@@ -23,13 +25,4 @@ public class FlowStep {
 
     @Field(type = FieldType.Integer)
     private Integer totalSteps;
-
-    public FlowStep() {}
-
-    public FlowStep(String flowStepId, String orderId, Integer activeStep, Integer totalSteps) {
-        this.flowStepId = flowStepId;
-        this.orderId = orderId;
-        this.activeStep = activeStep;
-        this.totalSteps = totalSteps;
-    }
 }
